# MAINT: stats.truncnorm.rvs: improve performance 