#### Reference issue
Closes gh-12370

#### What does this implement/fix?
gh-12370 reported that `truncnorm.rvs` slowed down by a factor of ~100 between 1.2.1 and 1.5.0rc2 in cases involving many different shape parameters. This PR improves the speed on that test case considerably:
In main, `533 ms ± 38.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)`.
In this PR, `4.94 ms ± 75.5 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)`

For a test case with only one set of shape parameters, `stats.truncnorm(13, 15).rvs(size=900)`, this PR is little slower:
In main, `1.79 ms ± 139 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)`
In this PR, `3.16 ms ± 171 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)`

The changes in `ppf` speed are likely quite similar, as this PR really just replaces the `_ppf` method and eliminates the custom inverse transform implementation of `_rvs`.

Hopefully, that's a fair compromise to close the issue, especially considering that the new `scipy.stats.sampling` will be much better than either version at sampling with a fixed set of shape parameters.

#### Additional information
The approach follows along the lines of gh-16586.